changed datatype from NUMERIC to FLOAT64 because NUMERIC was expected bigger bytesize than was given

+++ pipes/transformed-model-predictions/predictions_bigquery.py
@@ -21,7 +21,7 @@ def load_prediction_to_bigquery(weather_prediction: float, used_data_datetime: i
 
     # Define schema layout of BigQuery table.
     schema = [
-        bigquery.SchemaField("temp_pred", "NUMERIC"),
+        bigquery.SchemaField("temp_pred", "FLOAT64"),
         bigquery.SchemaField("prediction_datetime", "TIMESTAMP")
     ]
